Bug 1953740 - part 8 - Cleanup - observeAsComposableState can only return non-nullable r=android-reviewers,pollymce
This method returns the result of it's `map: (S) -> R` argument which returns
a non-null type so it will never return a nullable type.
Officializing this in the return type of the method allows avoiding the need
to null check the return in all places where this method is used.

This just updates the signature of the `observeAsComposableState` method.
Using the new signature in all places is to happen separately.

Differential Revision: https://phabricator.services.mozilla.com/D250803

UltraBlame original commit: c975671fbbe093e66d86111eb227842cb57ff4aa