Bug 1973524 - Fix internal crash in NavControllerImpl r=android-reviewers,gmalekpour,twhite
The crash is a regression from https://phabricator.services.mozilla.com/D249701. The intention there was to insure only once instance of `UnlockPrivateTabsFragment` at a time, but the combination of flags led to rare and difficult to reproduce crashes in https://cs.android.com/androidx/platform/frameworks/support/+/androidx-main:navigation/navigation-runtime/src/commonMain/kotlin/androidx/navigation/internal/NavControllerImpl.kt;l=585?q=NavControllerImpl. This patch aims to achive the same one-instance-at-a-time alighning it with approaches taken across the app, like in action_global_tabsTrayFragment or action_global_search_dialog.

What most likely is happening is, under rare timing conditions when navigating to `UnlockPrivateTabsFragment` with both `launchSingleTop=true` and `popUpToInclusive=false`, `NavController` skips the pop step (because it thinks the destination is already on top) and then skips the push step for the same reason. `FragmentManager` may still have that dialog attached, but `NavController`’s internal `backQueue` never actually removes the old entry. And on the next back-pop, the internal check in `popEntryFromBackStack` fails because the entry in the queue no longer matches the real top of the fragment stack (it appears that we are deeper than the distanation), resulting in an intermittent `IllegalStateException`.

By switching `popUpToInclusive` to true, we force `NavController` to always remove the existing `UnlockPrivateTabsFragment` entry before considering a re-navigation. That explicit removal keeps the internal `backQueue` in lock-step with `FragmentManager`’s actual fragments—old entries can’t linger behind—and therefore prevents the elusive out-of-sync state that was triggering the crash.

The downside of it, on repeated navigation calls the second instance might get created and immideatly destroyed, without reaching `onResume`. Well, better than a crash.

Differential Revision: https://phabricator.services.mozilla.com/D254902

UltraBlame original commit: 291acc35bd9300a0278e95a00ced2561fd6e50b3