Bug 1971558 - Schedule rendering the offscreen frame as soon as it is built. r=gfx-reviewers,jnicol
Previously the frame would be enqueued but only rendered when another frame was scheduled for rendering. It was correct because the only way to observe the result of an offscreen frame is to render something that reads from it but the new behavior is less surprising and less likely to cause jank.
Most of the content of this patch is making sure that offscreen frames don't contribute to the throttling mechanism.
I went the extra mile and plumbed the tracked flag from WebRenderAPI to the actual rendering so to make things as explicit as possible. It should be easier to keep the pending frame count balanced now that there is an explit flag for it.

Differential Revision: https://phabricator.services.mozilla.com/D253351

UltraBlame original commit: 09522a7b8c2d9db3d58952d260b6db75ecf6c2de