Bug 1951039 - part 1: Make `HTMLEditUtils::LineRequiresPaddingLineBreakToBeVisible` stop handling trailing collapsible white-spaces r=m_kato
The new white-space normalizer always makes the last white-space as
an NBSP if it should be visible but white-spaces are collapsible.
Therefore, the method won't be used for putting a padding `<br>` to
make the preceding collapsible white-space visible.  So, it does not
need to check the case anymore.

Differential Revision: https://phabricator.services.mozilla.com/D254472

UltraBlame original commit: 9b04407fbce082238d36dc272ecd8fc9a33eeadc