Bug 1953740 - part 2 - Extract functionality of configuring search suggestions providers r=android-reviewers,pollymce
Since AwesomeBarView contained all functionality for configuring
search suggestions providers and this logic would be needed in a
composable also the AwesomeBarView was renamed and cleaned to only
contain this logic.
A new class AwesomeBarViewSuggestions was introduced with the rest
of the configuration used by SearchDialogFragment for configuring
how to show search suggestions in a View.

Differential Revision: https://phabricator.services.mozilla.com/D250799

UltraBlame original commit: 1c40c52357315f7533e2eabea6e7a5c8544f3317