Bug 1959656 - Implement "Continuation API" feature for FedCM - r=anti-tracking-reviewers,webidl,smaug,timhuang
This patch implements the "Continuation API" as Chrome has branded it.
This feature provides the option for an identity provider to request a popup window to finish a login, rather than just providing an auth token.

To do this we had to do the following:
- Update the relevant parts of IdentityCredential.webidl
- Add IPC serialization of the new IdentityResolveOptions (IdentityCredentialSerializationHelpers.h)
- Refactor the resonse from the identity assertion endpoint to be a `IdentityAssertionResponse` rather than just a token (IdentityNetworkHelpers.{cpp,h}
- Create the IdentityProvider object, including its functions that allow the popup to communicate to the opener via the browser (IdentityProvider.{cpp,h})
- Add IPCs OpenContinuationWindow (to tell the caller it should open a popup) and ResolveContinuationWindow (to let the popup provide the token to its opener) (PWebIdentity.ipdl and associated "WebIdentity" classes)
- Flesh out the IdentityCredentialRequestManager to facilitate the communication of the token from popup to opener (IdentityCredentialRequestManager.{cpp,h}
- Tests, though not particularly thorough because we will have WPT coverage before we ship. (dom/credentialmanagement/identity/tests/mochitest)

The core logic lives in WebIdentityHandler (for sending responses to the RP from the IDP popup), WebIdentityChild (for opening the popup at the main process' request), and WebIdentityParent (for deciding when to open the popup, handling the eventual response, and maintaining IDP-RP linkage state). The rest is really a ton of plumbing.

A little cleanup slipped along with:
- Making the "tokens" returned UTF8Strings, rather than USVStrings
- The tokens stored on the IdentityCredential should not be Auto strings
- We don't "PromptUserWithPolicy" anymore, so this was renamed to "LinkAccount" and some dead code was removed

Differential Revision: https://phabricator.services.mozilla.com/D254252

UltraBlame original commit: ecb8ea5bf2dd1a6bdaf3cbcf227722e131b30fe6