Bug 1973288 [wpt PR 53297] - [soft navs] Include image element src changes as DOM modifications, a=testonly
Automatic update from web-platform-tests
[soft navs] Include image element src changes as DOM modifications

This CL changes soft nav detection so that changing the src attribute of
an HTMLImageElement --- as well as other attributes that cause an
image to be updated, e.g. src for <input type=image> or the poster
attribute for videos --- causes (future paints of) the node to be
tracked by soft nav heuristics. This is accomplished by:

 1. Moving the ContainerNode soft-nav checks into
    SoftNavigationHeuristics so we don't duplicate things, which is
    nice on its own to move decision making into SNH.
 2. Informing SNH of the attribute change, which we can do centrally in
    ImageLoader.

For now, this just reuses the ModifiedDOM() method to wire things up,
but we might want to add "cause" or something to add to tracing, which
can be done as a follow-up. We'll probably also want to consider other
attributes, but this is a starting point.

Bug: 328783345
Change-Id: I805386b4952627da6804e53bbdb9b114f4ba4e2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6599236
Commit-Queue: Scott Haseley <shaseleychromium.org>
Reviewed-by: Michal Mocny <mmocnychromium.org>
Cr-Commit-Position: refs/heads/main{#1476660}

--

wpt-commits: 0517b44d216fc3de60d300eae7684e89e4342170
wpt-pr: 53297

UltraBlame original commit: b5b0609df8ab4ce00714c882b2be6a241fd57ad8