Bug 1974096 - Create backdrop-filter root for [will-change] opacity r=emilio,tnikkel
Ensure that if opacity display item is created, which doesn't
result in an opacity surface (due to various optimizations)
that we still establish a backdrop-filter root as required.

Fixing this made the test which checks for backdrop-filter roots
with view-transition names fail, so fix that at the same time.

Differential Revision: https://phabricator.services.mozilla.com/D255034

UltraBlame original commit: 15f7d8c0e58f457c22842c731b0827f0797df3c8