Bug 1970339, part 1 - Text Fragments: Change API to create text fragments to use `Selection` instead of a single range. r=webidl,smaug,vhilla,mccr8
Before this patch, the C++ code was responsible
for creating a single text directive from a single
range. Accumulating the text directives into a
text fragment was done in the JS code.

This patch moves the responsibility of creating
the text fragment to the C++ code, which simplifies
the JS code, and improves the performance.

This is a prerequisite for the second patch, which
will make the timeout for creating text fragments
global for all ranges, instead of per range.

Differential Revision: https://phabricator.services.mozilla.com/D252945

UltraBlame original commit: c697f2d24243672a4e6f41150b4b6fc3cd160270