Bug 1961386 - part 2: Add MOZ_GUARDED_BY to mTimeout in nsTimerImpl. r=xpcom-reviewers,nika,jlink
Given TimerThread::Entry contains a copy of the timeout, we can use that everywhere we have no lock on nsTimerImpl::mMutex, relying on the invariant that the timeout must not change while the timer is in the TimerThread.
Doing so enables us to annotate it and have stricter checks.
Note that one unsafe, DEBUG-only check remains but it is explicitly annotated as such. TSAN will tell if we are safe enough.

Differential Revision: https://phabricator.services.mozilla.com/D249898

UltraBlame original commit: 26126f0db5a0cd965dd2d8176d4fec5899899871