Bug 1974302 - part 2 - Use the RegistrableDomainSpan for highlighting the domain r=android-reviewers,pollymce
Ensure integrators of the composable toolbar correctly mark the indexes range
of the domain in the URL to display and that then the composable toolbar will
use this data to correctly highlight the domain.

By leveraging a SpannableString as the URL it can be cut (like it happens
when Fenix wants to display a shorter version of the URL) while still ensuring
the integrity of the domain indexes marked previously marked in the SpannableString.

Differential Revision: https://phabricator.services.mozilla.com/D255433

UltraBlame original commit: f9175728f6ba42d1edd3de0b965382fa4c6ec850