Bug 1965846 - Use DoSetRange instead of SetStartAndEnd in CrossShadowBoundaryRange::ContentWillBeRemoved r=smaug
This is an existing bug that was revealed once the flat-tree copy-pasting
behavior was implemented.

Basically `SetStartAndEnd` does many extra things like comparing the
points so that it'll set the offset of the new boundaries accordingly.
This is wrong for `ContentWillBeRemoved` because the new offset could
soon become invalid as the content is being removed. We should just
call `DoSetRange` here and this is also what nsRange does.

Differential Revision: https://phabricator.services.mozilla.com/D255538

UltraBlame original commit: 6282c0b707eb5fce403b1144cdf7aad1efce11eb