Bug 1964600 - Remove the crash helper PID from the IPC connectors r=afranchuk
This paves the way for the daemonization of the crash helper. First of all the
main process won't know the PID of the crash helper process directly, as it
won't be a child of the main process anymore. Additionally, child processes'
IPC channels will be created in the main process, and then both ends will be
handed to the crash helper and child respectively. Because of this it won't be
possible for a child to fetch the PID of the other endpoint of the pipe (all
operating systems record the process identifier of the process that *created*
the pipe as the endpoint, regardless of what process one endpoint is handed
to afterwards).

This patch removes all references to the crash helper PID and ways to fetch it
from the crash helper IPC machinery. See the following patches for how we
deal with authorizing certain operations now that we don't have the PIDs
anymore.

Differential Revision: https://phabricator.services.mozilla.com/D251558

UltraBlame original commit: e792d735168b80e7c7cf17bb91ee0f4bb769dca5