Bug 1975821 - Avoid SetDelay in Http3Session. r=necko-reviewers,kershaw,mxinden
We do not use SetDelay here as this can lead to an unexpected
immediate firing instead of rescheduling if the timer event was already
in flight but not yet executed. InitWithCallback will instead cancel also
in-flight-events. There is no significant difference in overhead.

Differential Revision: https://phabricator.services.mozilla.com/D256168

UltraBlame original commit: ef7753355933a507749136f6281301e4946ca925