Bug 1974103 - Allow Merino requests to be sent over OHTTP. r=daisuke,nanj
This adds two prefs/variables for Merino OHTTP config and relay URLs. When OHTTP
is enabled, `MerinoClient` calls both the config and relay URLs with every
fetch. There are a few other features in the browser that use OHTTP, and they
work similarly.

The `MerinoClient` constructor now takes an `allowOhttp` param that defaults to
false. When true and the prefs are defined, the client will use OHTTP. The main
Suggest Merino client passes true for `allowOhttp`, and the other clients that
rely on IP-based geolocation (weather, geolocation) use the default false value.

There are other ways the `MerinoClient` interface could work, but it's hard to
say which is best, so I went with this. For example, `MerinoClient.fetch()`
could take `allowOhttp` instead of the constructor; or instead of `allowOhttp`,
consumers could pass in the config and relay URLs directly, which would let
different consumers use different config and relay URLs.

Differential Revision: https://phabricator.services.mozilla.com/D256732

UltraBlame original commit: fd8beff599459f671492ae123cb3dfe1345f039b