Bug 1972804 - Restore REALM_INDEPENDENT flag after generator resume. r=jandem
The realmIndependentJitCode check in emitEnterGeneratorCode was checking the
self-hosted script that is setting up the generator code's frame (such as
GeneratorNext), instead of the generator code's script like it should be
checking. We were also redundantly setting the interpreterScript_ field when
running in the interpreter. It is faster and easier just to always set the
interpreterScript_ field.

The generator code needs to set the REALM_INDEPENDENT flag on the frame after
resuming execution. emit_AfterYield generates the code that executes
immediately after resume and references the generator's script.

Differential Revision: https://phabricator.services.mozilla.com/D255382

UltraBlame original commit: 811b012bc964b764e573b91bf44b2045d4f75a61