Bug 1976653 - Support multiply composition op for accelerated canvas. r=aosmond
Accept when the destination target is opaque alpha, the multiply op must be broken
up into multiple stages. This requires multiple iterations of rendering the source
with different blend modes to approximate it. It is still expedient to do this for
performance instead of falling back to software.

Differential Revision: https://phabricator.services.mozilla.com/D256944

UltraBlame original commit: fd40d846e667822732d286a3dbe680fa2cfbd99e