Bug 1974682 - Don't create HTML element accessibles. r=morgan
In bug 1825611 and 1828373 we added more logic to create accessibles
when elements have overflow styling. There is an edge case where HTML
element will recieve overflow styling, in that case we do not want an
accessible as we rely on the scroll offsets of the root viewport.

Differential Revision: https://phabricator.services.mozilla.com/D256329

UltraBlame original commit: 482fcf2ecd12a5511f3dfdb86d73f4a865918b48