Bug 1966812 - Cache rect bounds in DocAccessibleParent. r=Jamie
Moved GetCaretRect to HyperTextAccessibleBase. I could see this living
in TextLeafPoint, but it is a bit confusing when a point is a caret and
when it is a simple offset. I think we would need to either add an
mIsCaret field to the class, or have a special static CaretBounds
method.

Differential Revision: https://phabricator.services.mozilla.com/D256377

UltraBlame original commit: 5f34c29bdb1f6e8c91c50bb111ebdf139972879b