Bug 1974445 - Linux sandbox: genericize the socketcall unpack trap. r=gerard-majax
The sandbox has code to handle `socketcall`, a multiplexer for
socket-related syscalls used in Linux's oldest target architectures, by
unpacking the arguments and resubmitting the call using the individual
syscalls added in kernel 4.3 if available.  This allows seccomp-bpf to
filter the syscall arguments, and not just which call it is (see also
bug 1066750).

Currently this is a special case for `socketpair` and `getsockopt`; this
patch generalizes it and adds a number of other socket calls, which will
be used by the next patch.

Differential Revision: https://phabricator.services.mozilla.com/D255741

UltraBlame original commit: af0dbabbd05d6d78ee82929ea8e597ce3b999e09