Bug 1963664 - Only re-enroll in rollouts if they unenrolled due to bucketing or targeting r=nimbus-reviewers,emcminn
We previously only re-enrolled if the rollout unenrolled for any reason
*except* for opt outs. We also want to exclude users that cause
a pref change as that is also an opt-out signal.

However, most of the remaining unenroll reasons are due to recipe issues
and we shouldn't attempt re-enrollment in those cases. The only other
remaining reason is opting out of studies, which should not prevent
rollouts from re-applying when studies are subsequently re-enabled.

Differential Revision: https://phabricator.services.mozilla.com/D257516

UltraBlame original commit: 426d7cc7387b26393eb3455444a5583d0ac0e16b