Bug 1977797 - Fix verifyTheExtensionInstallationTest UI test r=aaronmt
The UI test was flaky after installing the "Bitwarden Password Manager" extension.

This was happening because the title of the extension install prompt wasn't properly verified because it was trying to search and verify "Bitwarden" but in reality "Bitwarden Password Manager" was displayed.

To fix this inconsistency I've added a check that will update the extension name to use the entire title when verifying the install extension prompt.

The UI test successfully passed 100x on Firebase ✅

Differential Revision: https://phabricator.services.mozilla.com/D257981

UltraBlame original commit: fc97cf881d12beb39e38130f4133e23dd4a714d4