Bug 1973033 - Part 12: Fix small allocation sizes so the full expected range is used r=sfink
The problem is that the size class used for the largest small allocaton is
currently the same as that used for the smallest medium allocation, when these
should be different.

The patch arranges this by bumping up medium allocation size classes by one.
Some refactoring is done to pass around a SizeKind to indicate when small or
medium allocations are in play. The important parts are in
BufferAllocator::SizeClassForAlloc, SizeClassForFreeRegion and SizeClassBytes.

One wrinkle is what the maximum size of small allocations should be. This must
be smaller than the smallest possible medium allocation. If we make it too
close though our encoded size representation will round it up to be the same.

Differential Revision: https://phabricator.services.mozilla.com/D255670

UltraBlame original commit: 217cb173aefb510debdf94d3fa2bf1a20dfe4998