Bug 1972649 - Add fast-path for Acceleralated Canvas2D to WebGL copies. r=aosmond
This mainly adds a path to trade gl::SharedSurfaces's between AC2D and WebGL, in
both directions, so that copies from AC2D to WebGL and WebGL to AC2D can avoid
readbacks. The SharedSurface can be exported from and imported to the separate
WebGLContexts.

This requires substantial refactorings of GLBlitHelper to support alpha conversion,
y-flips, and cropping more generally. SharedSurfaces can be converted to
SurfaceDescriptors, of which GLBlitHelper needs to support these extra options
to fully represent the export and import requirements for AC2D and WebGL.

TexUnpackBlob also needs to be restructured to support falling back in case the
SharedSurface can not be exported to a SurfaceDescriptor.

Differential Revision: https://phabricator.services.mozilla.com/D257959

UltraBlame original commit: c11e7594c36d9f058fcb1c7371a416015cd1cc8f