Bug 1615206 - [devtools] Make the "Group similar message" setting control repeating messages. r=devtools-reviewers,ochameau
The existing setting was added to control grouping similar warning messages.
Due to the wording of the setting, users were also thinking it could control
repeated messages (the small badge/bubble we put next to messages when successibe
messages have the same text).
This patch renames the warning group pref into something that can drive both of
these, and we make the repeat mechanism check the value of the pref befor repeating.

Note that for repeated message, unlike warning groups, this doesn't "ungroup"
the messages, but that should be fine like this.

A node test case is added, as well as a mochitest (event though the test was already
checking repeated messages in a specific context, it's still nice to have
some basic actual coverage)

Differential Revision: https://phabricator.services.mozilla.com/D257841

UltraBlame original commit: 0f56ee13b4b524e775a231bc06522357578a28e7