Bug 1976737 - Show both a toolbar divider and a progress bar based on toolbar gravity r=android-reviewers,skhan
The browser toolbar just needs to know where it is placed on the screen
depending on which it will know where to place a divider or a progress bar.

The divider will always be shown to ensure a clearer separation of features
making it obvious when separate UI elements from different features are
placed immediately below or above the toolbar.

Differential Revision: https://phabricator.services.mozilla.com/D257108

UltraBlame original commit: 5457157cb9e79887584aff294650d2ad4dc617b9