Bug 1976480 - Text Fragments: Don't extend the selection if only punctuation is selected. r=dom-core,avandolder,mccr8
This patch adds a new check in the beginning of the algorithm which leaves the
input range unchanged if its content is only punctuation characters.
This edge case is handled by checking if all characters are punctuation and would return early, hence not going into the algorithm that extends the range to word boundaries
(which would become more complex trying to achieve the same).

Additionally, this patch contains a minor refactoring: Other edge cases in the input range
(empty range, whitespace-only ranges) are moved into the function that extends the
range to word boundaries.

Differential Revision: https://phabricator.services.mozilla.com/D258246

UltraBlame original commit: 79170f140f14b8cc4b668e6137a623118ddc49cf