Bug 1933787 - Test traced values with JSObjectsTestUtils r=ochameau,devtools-reviewers,profiler-reviewers,canaltinova
This does not represent the finalized form for a lot of the entries in the
snapshot since some types are still not handled, but this at least lets us
exercise the code paths and ensure future additions don't change the
representations of any existing covered types. I would also like to move
the serialization code here into a module we can include from Firefox and
the profiler frontend but that's for a future patch.

Differential Revision: https://phabricator.services.mozilla.com/D253219

UltraBlame original commit: 0c7eeec56eab2c8953b1c6c85e767e0ea6f52355