Bug 1978897 - do not try to pin dragged tab groups r=nsharpley,tabbrowser-reviewers
This patch will not offer the drag-to-pin indicator when dragging a tab group near the start of the tab strip. This will also skip any attempts to pin or unpin a tab group on drop.

We have a logical invariant that says pinned tabs can't be in tab groups and  tab groups can't be pinned (unless we complete bug 1916050).

Right now, the drag-to-pin indicator shows up when dragging a tab group to the start of the tab strip. On drop, the code will attempt to pin the tab group, but that will fail due to protections in Tabbrowser. The result is OK, but it's better to not confuse the user and it's better to not try invalid operations.

Differential Revision: https://phabricator.services.mozilla.com/D258483

UltraBlame original commit: 2581c020014fb0c192f8aa3052ef15dc254516ee