Bug 1979112 - Try to clean up mipmap state before blitting. r=aosmond
If a texture is determined mipmap incomplete, this may prevent it from being
attached to a framebuffer for drawing. Depending on GLES or Desktop GL, there
are different criteria for mipmap completeness, so just try to clean it all.

Differential Revision: https://phabricator.services.mozilla.com/D258645

UltraBlame original commit: 12bf685e70030752d8724ade6d210e6f68646c59