Bug 1977710 - Make `DependenciesPlugin` format Gradle errors for Treeherder "Failure Summary". r=adhingra,android-reviewers,jonalmeida
This makes `android-gradle-dependencies` errors much more easy to work
with.

It's surprisingly hard to format Gradle logging output: custom loggers
are *not* configuration cache compatible.  This is the best approach
that I can find, but it will lead to repeated messages in some
situations.  That's not the end of the world.  I've tried to make it
grep-able/Searchfox-able: the "[gradle:error]:" token is pretty
unique.

Differential Revision: https://phabricator.services.mozilla.com/D257606

UltraBlame original commit: f830ca8a3ad13346feca6663ac2f39b424823b47