Bug 1889757 Part 6: Pass messages from NativeLayerRootRemoteMacChild to NativeLayerRootRemoteMacParent. r=mstange,gfx-reviewers,nical
This handles the NativeLayerRoot methods to turn them into messages sent
to the NativeLayerRootRemoteMacParent. The layers are identified by
their GPU-side (client) pointers, which are turned into flavorless u64
values. When layers are created, we send creation messages noting the
size and other params of the layer creation. Later, when we
CommitToScreen, we pull properties from each of the layers, not knowing
which of those properties might have changed. Together, it should be
enough for the parent to reconstruct its layers and whatever new
properties they should have, including the IOSurfaceRef IDs.

The messages are queued up in a NativeLayerCommandQueue object, which is
created by the NativeLayerRootRemoteMacChild and then shared with all of
the NativeLayerRemoteMac objects it creates. The layer objects are
responsible for posting LayerInfo and LayerDestroyed messages; all other
messages are the responsibility of the NativeLayerRootRemoteMacChild.

This also modifies SurfacePoolCA to make calls to IOSurfaceCreate supply
the kIOSurfaceIsGlobal property, which is necessary for the surface to
be accessible by the parent process. This is a deprecated mechanism that
should be replaced, which will be done as a follow-up.

Differential Revision: https://phabricator.services.mozilla.com/D251183

UltraBlame original commit: 58524d6fab76ed973553d523e462ed389f49fde9