Bug 1889757 Part 2: Define a NativeLayerCA variant that receive an IOSurfaceRef directly. r=mstange
This adds yet another possible surface source for NativeLayerCA: a
surface that is handed to it from some other entity. For this patch
series, it will be receiving surfaces from the
NativeLayerRootRemoteMacParent, after it has looked up the surface from the
ID it received from the NativeLayerRootRemoteMacChild.

This also incidentally removes the unused NativeLayerCA::NextSurface
method. That method was made redundant in the refactoring of Bug
1962217.

Differential Revision: https://phabricator.services.mozilla.com/D251968

UltraBlame original commit: 4d4289d8006088904c4632176b404e981d33787b