Bug 1980170 - Fix the missing "Machine Learning" preset description r=mstange
It looks like when this was added, the preset description l10n key had a
trailing `2`. This is most like a copy/paste error.

We can't really change the l10n key inside the ftl file since it will
remove all the existing localizations in Pontoon, that's why we change
the name in the place where we use it.

Differential Revision: https://phabricator.services.mozilla.com/D259247

UltraBlame original commit: f32973f7ff3cd76bf5fa235184b3944f5af0493c