Bug 1979290 - Move testdriver implementation into TestDriverProtocol, r=Sasha
Instead of using multiple inheritance and a mixin class on the
Executor classes, follow the pattern of the rest of the code by using
the TestDriverProtocolPart to encapsulate protocol-specific
implementation of the testdriver control loop.

In particular this adds a run() method to the TestDriverProtcolPart
which is responsible for running the testdriver loop until there's a
test result. It defers to a get_next_message method for reading the
next message from the browser.

Differential Revision: https://phabricator.services.mozilla.com/D258688

UltraBlame original commit: d70564328caabea3706e28456dba5df1cf599c57