Bug 1979477 - Implement "New" badge logic for the visual search/Google Lens context menu item. r=Standard8,fluent-reviewers,urlbar-reviewers,bolsson
This adds a helper to `nsContextMenu` that shows or hides the "New" badge for a
given menuitem, so we can use it for any item in the content context menu.

As with the Perplexity badge in D253104, I don't see a way around having to
manually fetch the "New" l10n string from Fluent. It would be nice to be able to
use a string with the `.badge` attribute instead, but there's no way to do that
without duplicating the main menuitem string ("Search Image") so that there's a
version with the attribute and a version without the attribute. And that doesn't
scale anyway. It would be nice if there were a way to merge Fluent strings so
you could add attributes from one string to another.

I added `nsISearchEngine.isNewEngineOrURL()`, which can take an optional URL
arg, analogous to how the recently added `displayNameForURL()` can take an
optional URL. "isNewEngineOrURL" is more awkward as a name though. It might make
sense to split it into two functions, `isNewEngine()` and `isNewURL()`.

This does not update `search-config-v2-schema.json` because I'm planning to do
that when I vendor the related app-services change.

Depends on D258655

Differential Revision: https://phabricator.services.mozilla.com/D259182

UltraBlame original commit: 0ea34722a54a0b267a43c7fb5ddda31bc36b2f99