Bug 1974126 - Text Fragments: Allow users to copy an existing highlight. r=Gijs,webidl,emilio
Other browsers provide the "copy link to highlight" context menu entry  if the user right-clicks
onto an existing highlight (without selection). In this case, the created link contains all highlights
on the page. If a selection exists as well, it should take precedence.

This patch adapts to this behavior. When the context menu is opened, it is determined
whether there is a selection or an existing highlight range at the given point.
This has the side-effect that the "Copy Link to Highlight" option only shows up
if the user right-clicks on a selection or an existing highlight, but not anywhere else
(this also aligns behavior to Chrome).

Differential Revision: https://phabricator.services.mozilla.com/D259730

UltraBlame original commit: 0f96bad5ae6f15544d8ef79f3609746f42a8de76