Bug 1981471 - Part 1: Fix GCHashMap::traceWeakEntries so it uses its argument |e| rather than creating an Enum r=jandem
This is a problem because destroying the Enum may require taking the store
buffer lock which we don't hold at this point. That's why these methods take an
Enum, so that we can control its lifetime like this.

This brings it into line with the GCHasSet version.

Differential Revision: https://phabricator.services.mozilla.com/D260172

UltraBlame original commit: a4ac153694bbb31ae6c11045479014f651ac3951