Bug 1932700 - Fix edge case: click events inside buttons with popover descendants r=dom-core,smaug
Given a structure like:

```
<button popovertarget=foo>Activate
  <div popover id=foo>Clicking me shouldn't close me</div>
</button>
```

With the popover open, clicking the popover will propagate a click
upward to the button, which will close the button. This change ensures
that the original target is checked, and if it is a
ShadowIncludingDescendantOf the button then the popover behaviour is not
invoked.

Differential Revision: https://phabricator.services.mozilla.com/D259955

UltraBlame original commit: c966ce3fbecccc0f8d0f08f373a600d71d9a2f7a