Bug 1981339 - Set up menuitem icon if needed on image insertion. r=mac-reviewers,mstange
This popup is a bit weird because it's recreated from scratch every time
it's shown, so the items don't have the <img> element by the time we
create the native menuitem.

Watch for the <img> insertion and update the icon then.

Will file a follow-up to avoid recreating the menu over and over.

Differential Revision: https://phabricator.services.mozilla.com/D260119

UltraBlame original commit: 94782f60d41dbf440d8e8894845419a231280c04