Bug 1933726 - Cancel MainScope in IntentReceiverActivity to prevent memory leak. r=tthibaud,android-reviewers
IntentReceiverActivity uses `MainScope()` but did not cancel it, which
could lead to a memory leak if the activity is destroyed while the
coroutine is still running. This patch adds logic to store the scope in
a field and cancel it with `onDestroy()` to prevent this.

Differential Revision: https://phabricator.services.mozilla.com/D259989

UltraBlame original commit: 6324b86b582cbe9e42fc938029a0d2ec17b2ef04