Bug 1981327 - part 2 - Also switch off expanded toolbar layout when toolbar redesign is disabled r=android-reviewers,skhan
We wanted to keep the number of conditions to evaluate for knowing when the
navbar is enabled small so the current code only checks for when the expanded
toolbar layout is enabled or not.
This would allow for a bug in which the redesign with expanded layout is
enabled and then the redesign is disabled but the extended layout preference
still remains leading to the application thinking that it should show the
redesigned toolbar with expanded layout.

Differential Revision: https://phabricator.services.mozilla.com/D260104

UltraBlame original commit: 8282dcd978b70d462ea9cd6fe05c0cdd8cfb7573