Bug 1972460 — add functioning connection time for ip protection panel. r=ip-protection-reviewers,fluent-reviewers,bolsson,fchasen
This patch adds a time connected string to the ip protection status card in the panel. The string only appears when protection is enabled and is based on the property protectionEnabledSince. It uses an interval to update the string every second. The interval should stop when the panel closes or when protection is disabled.

Differential Revision: https://phabricator.services.mozilla.com/D256738

UltraBlame original commit: f36d95f3920344b9ccc84157d6920d3fbf6e4c58