Bug 1980613 - Part 11: Scalar replace MTypedArraySet. r=jandem
There are three cases to consider:

1. `subarray` is the target object in `set`: `ta.subarray(...).set(...)`

In this case we just need to adjust the `set` offset parameter to include
the `subarray` start parameter.

2. `subarray` is the source object in `set`: ta.set(src.subarray(...), ...)`

This case requires a new instruction to pass through the `subarray` start and
length operands.

3. `subarray` is both the target and source object.

This is case is a no-op instruction.

Differential Revision: https://phabricator.services.mozilla.com/D259587

UltraBlame original commit: 0d91709aabe92688fe1c901ed036d1b3fd874799