Bug 1982293 - Look up the pseudo-element when retrieving the starting-style rules for DevTools. r=layout-reviewers,emilio
This is the missing part when working on this API. We should find the
pseudo-element we expected (just like what we do for
`Element.getAnimations()`), and use this pseudo-element to resolve the
starting style.

The test is in https://phabricator.services.mozilla.com/D260796.

Differential Revision: https://phabricator.services.mozilla.com/D260907

UltraBlame original commit: 3a40b43391a877f79476284956bc6ff93f2f0bc1