Bug 1977726 — Add error message and state to the IP Protection panel r=ip-protection-reviewers,fluent-reviewers,rking,bolsson
This patch adds a new "ipprotection-message-bar" component based on "moz-message-bar" that renders messages specific to the IP Protection panel.

It also adds a new state property "error" to indicate the type of error to display on the panel. All supported types will be listed as constants in our constants file. The only message type available at this time is a generic error. We will render a message bar in "ipprotection-content" via a "_showMessageBar" lit reactive property. "_showMessageBar" is true if an error is detected in the state. Otherwise, no message bar is displayed by default with "_showMessageBar" set to false.

In the case the message bar is dismissed with the close button, we dispatch a separate event called "ipprotection-message-bar:user-dismissed" to any consumer element, like "ipprotection-content".

Differential Revision: https://phabricator.services.mozilla.com/D260538

UltraBlame original commit: 22c4c4bb4d0d478bddad6f1e9321cc0a22af0f8a