Bug 1980636 - Make `EventStateManager::UpdateCursor` stop updating the cursor when the trigger event is not for the last pointer r=edgar,dom-core,layout-reviewers,hiro
Now, synthesized `ePointerMove`s are fired for all extant pointers if
either a scroll or a reflow occurs.  Then,
`EventStateManager::PreHandleEvent` updata cursor for all synthesized
`eMouseMove` and `ePointerMove`.  However, cursor should be managed with
the last pointer.  Therefore, most synthesized `ePointerMove` is not
for the proper pointer to update the cursor with its target.

This patch makes `EventStateManager::UpdateCursor` ignore synthesized
events which are not for the last pointer which the process handled the
last pointer.

Note that forgetting the last pointer when the last pointer leaves from
the documents managed by the process is important because the last
pointer may be moved into a remote content.  Then, no pointer in the
parent process shouldn't cause updating the cursor.

Differential Revision: https://phabricator.services.mozilla.com/D260916

UltraBlame original commit: a648b2ed7e3b10fbd641371362b79b55a3cc6a15