Bug 1982512 - xz-embedded owner shouldn't be GeckoView. r=application-update-reviewers,geckoview-reviewers,bytesized,ohall
At first, xz-embedded was imported for Fennec, but it is not used in
GeckoView now. Actually, this library is used in updater. So we would
like to move the ownership to updater since this library is used in
updater only.

Differential Revision: https://phabricator.services.mozilla.com/D260777

UltraBlame original commit: 625a5d0ef66d6518610018e802f31e7cb2c35ca3