Bug 1918874 - Use NS_DispatchToCurrentThread instead of NS_DispatchToMainThread in HttpChannelChild::Release to avoid shutdown leak r=necko-reviewers,kershaw
We are guaranteed to be on the main thread in HttpChannelChild::Release() as
we always dispatch to MT otherwise. So we can use NS_DispatchToCurrentThread
to enqueue the DoNotifyListener runnable, instead of NS_DispatchToMainThread
which throws an NS_ASSERTION during process shutdown and leaks the runnable.

Differential Revision: https://phabricator.services.mozilla.com/D261374

UltraBlame original commit: 42f4f33bf8eb603dd86dee7f2bbf7d812520a707