Bug 1981267 - Modifies AndroidImageDecoder to use scale with filtering r=android-reviewers,avirvara
Previously we just downsampled the bitmap to the desired size. With jpg encoded images this resulted in
blocky looking text when previews of a tab were presented on the home screen or in the tab tray. Now we downsampe
to the closest largest size we can and perform a filtered bitmap scale to the desired size. This will alias the text
during the scale increasing legibility.

Differential Revision: https://phabricator.services.mozilla.com/D261202

UltraBlame original commit: fa44232c6a505d4e64c7b32ab1bf0fe61c15b271