Bug 1981937 - [devtools] Don't clear stylesheets when selecting a document in Browser Toolbox iframe picker. r=devtools-reviewers,jdescottes.
In such case, we don't want to clear the stylesheet editor, as we're not calling
watch again for the target (since it didn't change), and We shouldn't clear the
StyleEditor list either.
The StyleEditor doesn't really adapt to the iframe picker document selections,
even in regular cases, so hopefully we're not making it any worse here.

Differential Revision: https://phabricator.services.mozilla.com/D260790

UltraBlame original commit: 6e48cfc728fa09f6795a17d770862bc70c317cad