Bug 1979218 - Add enrollment checks for IP Protection. r=ip-protection-reviewers,sstreich,omc-reviewers,pdahiya
This limits the use of the IP Protection widget to those who are in the experiment or have already enrolled, instead of relying on the browser.ipProtection.enabled pref.

Enrollment is a one time permanent link of the VPN feature to a users account when an eligible user turns on the VPN.

An enrolled account can access the VPN, even when their device is not part of an experiment, as long as`browser.ipProtection.enabled` is true.

- Adds the browser.ipProtection.enabled pref to the synced preferences.
- Inits the IPProtection widget if a user is already enrolled or the device is eligible to enroll.
- Listens for ipProtection nimbus updates in the IPProtectionService and enrolls users who are in an experiment when they are signed in.
- On FxA sign in, the IPProtectionService checks if a users account has been previously enrolled.
- Prevents starting the proxy if not enrolled.

With these changes, being in a ipProtection experiment is required to test the widget.

Differential Revision: https://phabricator.services.mozilla.com/D260401

UltraBlame original commit: 184f079fc25c6df6cd53e74ed90c67ff14305022