Bug 1984014 - Check for untransformed OP_SOURCE rectangles when drawing. r=aosmond
Sometimes we submit rectangles for drawing without a transform to bypass the transform.
However, for OP_SOURCE, we expected there to always be a transform to check when to clear
the background. This fixes the OP_SOURCE check to account for untransformed rects.

Differential Revision: https://phabricator.services.mozilla.com/D261805

UltraBlame original commit: 1c3f976e10d07382f96f0be6dfa3c7ed02ff7170