Bug 1984123 - Override WebGPU CTS' baseResourcePath for firefox's CTS vendoring scheme. r=webgpu-reviewers,ErichDonGubler
The baseResourcePath assumes tests are being run from the parent of
the CTS's resource directory. This is not the case with how we vendor
the CTS. As a result, attempts to load resources during the CTS
are failing with 404 errors.

This patches the CTS to set the default baseResourcePath to the
correct location. It also ensures that the overridden value when
running with workers is also correct.

We must run `patch` as a run-command action prior to the `cargo run`
step, as it patches the original typescript sources which are used as
an input to the rust step. If we were to use moz.yaml's `patches:`
mechanism, the typescript sources would be patched *after* they had
already been processed and the tests had been written to their final
locations.

This causes a bunch of tests in web_platform/copyToTexture/image_file
to now pass. Expected failure annotations are also removed from a lot
of other tests. Note, however, that these do not actually pass, but
are rather skipped due to HTMLVideoElement.captureStream not being
supported.

Differential Revision: https://phabricator.services.mozilla.com/D261763

UltraBlame original commit: 5324598601d01d3a938f8a1ac6eb06f0bec4f46b