Bug 1984981 - Update the layout for showing the clipboard banner r=android-reviewers,sfamisa
Investigations shown a race between the measure/layout phases of the
InlineAutocompleteEditText (View) and the rest of the composable layout
for showing search results.
Until being able to convert that View to Compose the best we can do is
build our layout in such a way to avoid the issue. Since the suggestions
content is already shown in a Column we can avoid the new Column
composable which shows the issue fixed.

Differential Revision: https://phabricator.services.mozilla.com/D262337

UltraBlame original commit: 7c3fbc4aac664fbd6722a63a3de2b37d743f9c1d