Bug 1984500 - Remove separate dom.ipc.processPrelaunch.startupDelayMs pref. r=nika
It was set to the same value as the regular dom.ipc.processPrelaunch.delayMs pref (1000ms).

Moreover, I'd like to change the startup preallocation to happen immediately and not
from a runnable that's posted to the main thread, so we won't be calling AllocateAfterDelay
for the startup case, and we can remove the argument.

Differential Revision: https://phabricator.services.mozilla.com/D262027

UltraBlame original commit: f26d0fd91b3724a7e4ef09071f2b3ff2560f3651