Bug 1980274 - Part 2: Add a button to mute audio when a Taskbar Tab starts playback. r=Jamie,nrishel,fluent-reviewers,accessibility-frontend-reviewers,desktop-theme-reviewers,sfoster,emilio
This button appears next to the 'reload' button and mutes audio when
clicked. It also appears pressed if the tab is muted to make it more
clear and try to provide an analogy to a checkbox.

This also adds strings to indicate what it will do when pressed.

Differential Revision: https://phabricator.services.mozilla.com/D260358

UltraBlame original commit: adc11dfad27242698d97090fa0ef1af05f81407f