Bug 1982440 - Avoid invalidating device cache twice r=pehrsons
The PipeWire camera backend is using a placeholder device, which is used
before the camera access is fully granted. Once the camera access is
granted, we go through the device list and invalidate the cache if we
find there is a placeholder device and enumerate camera devices again
to get the real ones. With the introduction of devicechange event, which
is emitted in VideoEngine::ClearVideoCaptureDeviceInfo() and called when
we the access is granted, we would end up invalidating the device cache
again. We can instead rely on this new devicechanged event and drop all
the "IsPlaceholder" property usage.

Differential Revision: https://phabricator.services.mozilla.com/D261088

UltraBlame original commit: 5bb82d8283c4a174a3a580433df5d11d51c8a47a