Bug 1984104 - Run success steps after committed promise has resolved. r=smaug,jjaschke
The spec requires there to be a committed to entry when the success
steps run, but that isn't set until #notify-about-the-committed-to-entry
runs. This means that we need to wait until there is a committed to
entry before running the success steps. The way we choose to perform
this is by waiting until the committed to promise resolves which allows
us to re-sync with how spec perceives believes things should look.

Differential Revision: https://phabricator.services.mozilla.com/D261750

UltraBlame original commit: d7e610ef696c00fd2ba8b682ffa0f74a63027cd2