Bug 1980753 - Fix AppLinksFeature not always reacting to the currently selected tab r=android-reviewers,Roger
Moving the feature to the onViewCreated ensures that we are always observing the selected tab without needing to use the `customTabSessionId` to specify the selected tab every time - this is something that caused side-effects.

By moving it here, we also avoid re-initialising the feature when we enter the BaseBrowserFragment from the various entry points (e.g. tabs tray and swipe gestures).

Differential Revision: https://phabricator.services.mozilla.com/D262596

UltraBlame original commit: 4c0fec03ec58131452b52d8b2fe359eee5479450