Bug 1985208 - Don't restart input if the edit toolbar loses focus r=android-reviewers,skhan
This was added in 2018 to fix a bug with back presses on the IME not
deleting the intended text -
https://github.com/mozilla-mobile/android-components/issues/1489
Recently this was seen to cause issues when using this functionality
in the composable toolbar - the keyboard has glitches in closing.
Testing various scenarios without this showed the original issue is
no more - back presses deletes the right text - and that the issue
seen in the composable toolbar disappears also.

Differential Revision: https://phabricator.services.mozilla.com/D262546

UltraBlame original commit: 2bb38758d15beb719e3a2677da0b66dace55a964