Bug 1980588 - Add PayloadFlags to the old and Rust marker API r=mstange,profiler-reviewers
This changes a lot of places, but it's essentially just find and
replace. We had to change this old marker API, because the new API
uses this one to serialize the marker schema fields.

This patch removes the AddKey{Label,}FormatSearchable methods and
switches to use the original AddKey{Label,}Format methods with an
optional parameter instead to make sure that we can pass multiple flags
instead of just Searchable.

For Rust, since we can't have default argument values, we had to create
a new method for them.

Differential Revision: https://phabricator.services.mozilla.com/D260465

UltraBlame original commit: 7de5200acac548d5b094e96e7eb1d00d95ac426f