Bug 1985382 - Reorganize PerfSpewer to split the debug entries and IR file generation. r=iain
PerfSpewer entagles two separate concepts:
  1. The JitDump debug entries that map from PC -> line/column
  2. The IR opcodes we generate when IONPERF=ir that are printed
     to a temp file and linked to from the debug entries.

This is a problem when trying to support printing the iongraph.json
to jitdump, as we don't want to use the IR opcode feature of PerfSpewer
but we do want to have debug entries that reference the JSON.

This commit refactors PerfSpewer so that it only holds a collection
of debug entries that will be written into JitDump later. The IR
file is now opened eagerly when we begin recording, and all writes
to the file also add to the debug entries vector.

This split makes IONPERF=src cleaner, as it only needs to add
debug entries. This will also simplify iongraph integration as
we can now just have the GraphSpewer write directly into the
file and avoid an extra buffer.

BaselineInterpreterPerfSpewer keeps a version of the Op struct
as it makes a late decision about whether to generate multiple
JitDump entries or just a single one for the loop, and it's
easiest to just have the IR in memory for this.

Differential Revision: https://phabricator.services.mozilla.com/D262674

UltraBlame original commit: c5f54b1b749ccfc03b1bbd390f53fb07c6db5ebc