Bug 1983601 - Fix recording app opens in Nimbus r=android-reviewers,sfamisa,twhite
The bug was that we didn't show a review prompt at all when telemetry was disabled.
The root cause was that we require a certain number of app opens as one of necessary conditions to show a review prompt. We check it using Nimbus' behavioural targetting functionality (which works on-device). But we were only recording app opens in Nimbus when telemetry was enabled.
The fix is to move recording app opens out of a block that only runs when telemetry is disabled. While it makes sense for this to be close to telemetry that tracks the same thing and it (probably) makes sense to avoid some extra work related to the telemetry when we know we aren't going to send it, we don't need to skip recording app opens in a local event store.

Differential Revision: https://phabricator.services.mozilla.com/D263084

UltraBlame original commit: ad09f0ed642ef396d7cf4ff05982d305be583cea