Bug 1984004 - Ensure explicit popover invokers participate in focus fixups r=dom-core,smaug
nsFocusManager assumed that the invoker of a popover would always have
a `popovertarget` pointing to the popover, however this is not the case
as the invoker of a popover can be any element set via
`showPopover({source})`.

This change ensures that `showPopover({source})` sets a WeakPtr on
`source` to the popover - the `mAssociatedPopover` extended slot.

This also refactors nsFocusManager to use `GetAssociatedPopover()`
instead of looking up the element's popovertarget element.

Differential Revision: https://phabricator.services.mozilla.com/D263142

UltraBlame original commit: e6b416e8ad02affc9f5187497d0e17ccdfff21d6