Bug 1985654: Allow anchor elements to have inline elements as its 'containing block'. r=layout-reviewers,emilio
Technically, we are following the spec to the letter. However, now-passing WPTs
indicate a conflict. I think it fits the spirit of the spec (Where the guarantee
we want is that an anchor is fully laid out by the time we need to reference it)
better to allow inline elements as anchor elements' containing blocks.

Differential Revision: https://phabricator.services.mozilla.com/D263031

UltraBlame original commit: f0899e3894485bb8eb8bdef14751f51f04a8979f