Bug 1986472 - Ensure we can handle losing prepare races with GPUVideoTextureHost and SurfaceTextureHost. r=gfx-reviewers,lsalzman
When GPUVideoTextureHost is created, we just have a handle ID that is
used to bind to the actual underlying TextureHost from the decoding
process. This has an implicit race condition where content might
reference the ID before the decoding process creates the texture in the
compositing process.

This means that when GPUVideoTextureHost is used in conjunction with
SurfaceTextureHost, we can lose the race to call PrepareForUse on the
wrapping RenderAndroidSurfaceTextureHost object on the Render thread.
Without PrepareForUse, we cannot compositing the surface.

In order to resolve this, this patch makes it so that we call
PrepareForUse if it has not yet been called when we call NotifyForUse.
This allows the state machine to move forward gracefully. We already do
this for the RemoteTextureHost path.

Differential Revision: https://phabricator.services.mozilla.com/D263419

UltraBlame original commit: 62c373d81e2481642bef229d2b1f4ea00991f639