Bug 1986107 - Ensure we gracefully handle unsupported codecs in ChannelMediaDecoder::CreateStateMachine. r=media-playback-reviewers,karlt
While we should not be encountering an unsupported codec on this
pathway, we should gracefully handle it in the event it occurs. This
patch adds the necessary plumbing as well as logging so that when users
report playback failures, we can identify this situation more easily.

Differential Revision: https://phabricator.services.mozilla.com/D263762

UltraBlame original commit: b1c580ca658e99fa47ca18c393ac0be43a63c58f