Bug 1985600 part 5 - Change isStore check in MDefinition::dependency to an assertion. r=rhunt
`getAliasSet()` is a virtual method so this check can add quite a bit of overhead.

We used to have a `storeDependency_` field in the union with `loadDependency_` and that's
why the check was originally required, but the `storeDependency_` field was removed in
bug 1687255 part 16.

Differential Revision: https://phabricator.services.mozilla.com/D263152

UltraBlame original commit: 6e9bc8a859c4436148049bd223e084f8546ff8fd