Bug 1986421 - Use lit directives to mange connection time timer  r=ip-protection-reviewers,rking
This Patch changes the current conneciton time to be a generator instead of a prop.
This allows us to offload the "should the timer update" decision into lit-directives.
Which then hooks into whether the component is rendered, so we no longer need to manually start/stop the timer.

Differential Revision: https://phabricator.services.mozilla.com/D264071

UltraBlame original commit: 095b4c77c84ac2e04661120854278467883f6bec