Bug 1987209: Add browser.backup.profile-restoration-date when restoring backups r=profiles-reviewers,mossop
Pref should contain this timestamp (Math.round(Date.now/1000)) in the current
profile if and only if the current profile was obtained by a restore of a
backup.  This is achieved by barring the pref from ever being serialized, and
adding the pref to the end of prefs.js after deserialization.

Differential Revision: https://phabricator.services.mozilla.com/D264122

UltraBlame original commit: 4ab8d0e50cee660636e3394682945d2783149751