Bug 1930008, when performing transformations on credit card and address fields, always use the original value rather than trying to re-transform the modified value again, r=dimi,credential-management-reviewers
Most of this change is moving the transformers into a separate class ProfileTransformer. A special property _original is used to store the original values. It isn't a private or internal property so that the values are transfered over processes.

Differential Revision: https://phabricator.services.mozilla.com/D264471

UltraBlame original commit: 35c0731eba5f7178edb4eaff05938d2b4c1b0045