Bug 1990448 - Check ShouldSkipFrame on the out of flow. r=layout-reviewers,tnikkel
Placeholders don't have the right style. This fixes
-moz-subtree-hidden-only-visually on out of flows, which is currently
not used but might be needed for the front-end.

Unfortunately this is restricted currently to the parent process, which
means that my attempt at a reftest didn't work... I could write a
browser test by absolutely positioning a tab panel or something but that
seems a bit clunky...

Differential Revision: https://phabricator.services.mozilla.com/D265957

UltraBlame original commit: 29be892785c8a5013c5793387049c1966541ff61