Bug 1989107: Share generateEnterJit code across architectures r=jandem
To avoid having to rewrite ~7 arch-specific EnterJit stubs to remove the need for a rectifier frame, this patch adds a shared path for all architectures (except arm64; see next patch).

I initially tried to rewrite the calling convention to eliminate the awkward way that actualArgs is passed in the result value, under the assumption that it was always the case that actualArgs+1 == maxArgc. However, it turns out that's not true when doing OSR into blinterp. In that case, we already padded the arguments array with undefined when doing the initial call, so argc (the size of the argv array) is max(actuals, formals). However, when creating the frame descriptor, we need to use the original actualArgs value, because it's observable for things like arguments.length. So I rewrote the patch to reverse the removal of actualArgs. I left in my changes to pass argc/argv without `this`, because it simplified codegen. I added a testcase that would have caught my initial bug.

The only part of this trampoline that needs to be architecture-specific is the ABI calling convention. We could (and probably should) move more code into generateEnterJitShared, but this was a good dividing line for now that lets us eliminate the rectifier stub. I opened bug 1989118 to track the remaining work.

I didn't run the entire testsuite on the unsupported platforms (mips/loong/riscv), but I did verify that each platform compiles and can run basic tests.

Differential Revision: https://phabricator.services.mozilla.com/D265899

UltraBlame original commit: 838c8b7e375bb6b9d5436f692c9456ec64963f4e