Bug 1990729 - Don't #include flat_set.h when compiling libwebrtc as C++20. r=mjf,webrtc-reviewers
This change allows libwebrtc to be compiled as C++17 or C++20, but doesn't switch to compiling as C++20 yet.

After we do compile libwebrtc as C++20, we can remove third_party/libwebrtc/moz-patch-stack/s0122.patch, but in the meantime we still need it for C++17.

Differential Revision: https://phabricator.services.mozilla.com/D266114

UltraBlame original commit: aadccad67497469c79a9239725e26e0867e20277