Bug 1991290 - Apply overflow-clip-box to input's scroll containers explicitly. r=tnikkel,layout-reviewers
As per other discussions we don't really want to expose this property to
authors, so there's no benefit in letting them override it.

This property's content-box value doesn't play nice with border-radius
(well, it does deal with it correctly, but it doesn't do what we want it
to do).

Differential Revision: https://phabricator.services.mozilla.com/D266605

UltraBlame original commit: 55ced76236351abc8518b1534ee44d13b1de546f